int to double

+++ TestGroupProject-106/Class1.cs
@@ -13,7 +13,7 @@ namespace TestGroupProject_106
             return "Hello there!";
         }
 
-        public Class1(int i) { }
+        public Class1(double i) { }
         
     }
 }
